Refactor architecture, Update NuGet packages
