+++ VattenMedia.Core/Entities/ChannelId.cs
@@ -1,4 +1,4 @@
-﻿namespace VattenMedia.Common.Entities
+﻿namespace VattenMedia.Core.Entities
 {
     public class ChannelId
     {
