+++ VattenMedia.Core/Entities/Config.cs
@@ -1,4 +1,4 @@
-﻿namespace VattenMedia.Common.Entities
+﻿namespace VattenMedia.Core.Entities
 {
     public class Config
     {
