+++ VattenMedia.Core/Entities/Quality.cs
@@ -1,0 +1,9 @@
+﻿namespace VattenMedia.Core.Entities
+{
+    public enum Quality
+    {
+        High,
+        Medium,
+        Low
+    }
+}
