+++ VattenMedia.Core/Entities/QualityOptions.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace VattenMedia.Common.Entities
+namespace VattenMedia.Core.Entities
 {
     public static class QualityOptions
     {
