+++ VattenMedia.Core/Interfaces/IYoutubeService.cs
@@ -1,4 +1,4 @@
-﻿namespace VattenMedia.Infrastructure.Services
+﻿namespace VattenMedia.Core.Interfaces
 {
     public interface IYoutubeService : IStreamingService
     {
