+++ VattenMedia.Core/VattenMedia.Core.csproj
