+++ VattenMedia.Infrastructure/StatusManager.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using VattenMedia.Core.Interfaces;
 
 namespace VattenMedia.Infrastructure
 {
