+++ VattenMedia/Bootstrapper.cs
@@ -1,6 +1,7 @@
 ﻿using System.Configuration;
 using Unity;
 using Unity.Lifetime;
+using VattenMedia.Core.Interfaces;
 using VattenMedia.Infrastructure;
 using VattenMedia.Infrastructure.Services;
