--- VattenMedia/Models/Quality.cs
@@ -1,9 +1,0 @@
-﻿namespace VattenMedia.Models
-{
-    internal enum Quality
-    {
-        High,
-        Medium,
-        Low
-    }
-}
