+++ VattenMedia/packages.config
@@ -1,11 +1,13 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="RestSharp" version="106.6.3" targetFramework="net471" />
+  <package id="RestSharp" version="106.6.7" targetFramework="net471" />
   <package id="System.Configuration.ConfigurationManager" version="4.5.0" targetFramework="net471" />
   <package id="System.Drawing.Common" version="4.5.1" targetFramework="net471" />
   <package id="System.Reflection.Emit" version="4.3.0" targetFramework="net471" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.2" targetFramework="net471" />
   <package id="System.Security.AccessControl" version="4.5.0" targetFramework="net471" />
   <package id="System.Security.Permissions" version="4.5.0" targetFramework="net471" />
   <package id="System.Security.Principal.Windows" version="4.5.1" targetFramework="net471" />
-  <package id="Unity" version="5.8.13" targetFramework="net471" />
+  <package id="System.ValueTuple" version="4.5.0" targetFramework="net471" />
+  <package id="Unity" version="5.9.5" targetFramework="net471" />
 </packages>
